[TEST] API Test for Get Total Report Created At Month, Get Total Report Spending By Month, Get Total Report Checkout By Month

+++ gudangku/app/Http/Controllers/Api/StatsApi/Queries.php
@@ -684,7 +684,7 @@ class Queries extends Controller
                         'context' => Generator::generateMonthName($i,'short'),
                         'total_price' => $total_price,
                         'total_item' => $total_item,
-                        'average_price_per_item' => $total_item > 0 ? ceil($total_price / $total_item * 100) / 100 : 0
+                        'average_price_per_item' => $total_item > 0 ? (int)ceil($total_price / $total_item * 100) / 100 : 0
                     ]);
                 }
